% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{cms_kpi_rate}
\alias{cms_kpi_rate}
\title{Sales Commission Rates on bonus}
\usage{
cms_kpi_rate(total_prod, target, sales, cost)
}
\arguments{
\item{total_prod}{total of products a sale rep shall sell}

\item{target}{threshold to reach by the sales reps}

\item{sales}{achievement in term of products sold}

\item{cost}{mean of price and discount price}
}
\description{
Sales Commission Rates on bonus
}
\details{
Calculation of the bonus based on the performance of the sales reps. It depends to your commercial policies
}
\author{
Jean Marie Cimula
}

