% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{cms_markdown}
\alias{cms_markdown}
\title{Commission based on the Mark Down Percentage}
\usage{
cms_markdown(price, disc.price)
}
\arguments{
\item{price}{List price}

\item{disc.price}{Discount price}
}
\description{
Commission based on the Mark Down Percentage
}
\details{
Mark Down represents the price that is below the net selling price with which the product is sold to the customer. For each mark down value calculated, a % percentage will be retrieved from a predefined table with the
combination of above percentage and product hierarchy up to 4th level. It means that the mark down % in this
case will be the corresponding % value of 5 in the mark down table.
}
\examples{
List price = $2.00, Discount price = $1.90
cms_markdown(2,1.9)

}
\author{
Jean Marie Cimula
}

