% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{cmscalc}
\alias{cmscalc}
\title{The Calculation of Commission of sales reps}
\usage{
cmscalc(df, ..., com.type = c("Markdown", "Markup"), rmv.na = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{com.type}{type of the commission ( Markdown or Markup)}

\item{rmv.na}{boolean value of na (TRUE or FALSE)}
}
\description{
The Calculation of Commission of sales reps
}
\details{

}
\author{
Jean Marie Cimula
}

