% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{cms_tracker}
\alias{cms_tracker}
\title{Activation tracking of sales forces}
\usage{
cms_tracker(dataframe, ..., region = "ALL", from = FALSE, to = FALSE,
  set.target = c("50", "50", "50", "50", "50", "50"))
}
\arguments{
\item{region}{filter the region of activities. Default "ALL"}

\item{to}{the end of the period eg."2016-06-30"}

\item{set.target}{depends to the commercial policies in order to judge the achievement of salesforce}

\item{df}{data.frame}

\item{dt_from}{the beginning of the period eg."2016-06-01"}
}
\description{
Activation tracking of sales forces
}
\details{

}
\author{
Jean Marie Cimula
}

